주문 생성 로직 변경 및 레이어 간 의존성 변경 수정

+++ core/core-api/src/main/java/io/dodn/springboot/core/domain/product/response/ProductResponse.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 @Builder
 public record ProductResponse (
 		 Long id,
-		 int productNumber,
+		 Long productNumber,
 		 ProductType type,
 		 ProductSellingStatus sellingStatus,
 		 String name,
